Data Structures Fundamentals with C#/S04_Trees_Representation_and_Traversal_(BFS and DFS)_Exercise/Tree/Tree.cs
namespace Tree
{
    using System;
    using System.Collections.Generic;
    using System.Collections.ObjectModel;
    using System.Text;

    public class Tree<T> : IAbstractTree<T>
    {
        private readonly ICollection<Tree<T>> children;

        public Tree(T key, params Tree<T>[] _children)
        {
            Key = key;
            children = new List<Tree<T>>();

            foreach (var child in _children)
            {
                AddChild(child);
                child.Parent = this;
            }
        }

        public T Key { get; private set; }

        public Tree<T> Parent { get; private set; }

        public IReadOnlyCollection<Tree<T>> Children => (IReadOnlyCollection<Tree<T>>)children;

        public void AddChild(Tree<T> child)
        {
            children.Add(child);
        }

        public void AddParent(Tree<T> parent)
        {
            Parent = parent;
        }

        public string AsString()
        {
            var sb = new StringBuilder();

            DfsAsString(sb, this, 0);

            return sb.ToString().Trim();

        }

        private void DfsAsString(StringBuilder sb, Tree<T> tree, int depth)
        {
            sb.Append(' ', depth)
              .AppendLine(tree.Key.ToString());

            foreach (var child in tree.children)
            {
                DfsAsString(sb, child, depth + 2);
            }
        }

        public IEnumerable<T> GetInternalKeys()
        {
            throw new NotImplementedException();
        }

        public IEnumerable<T> GetLeafKeys()
        {
            var result = new List<T>();
            var queue = new Queue<Tree<T>>();

            queue.Enqueue(this);

            while (queue.Count > 0)
            {
                var curruentSubTree = queue.Dequeue();

                if (curruentSubTree.children.Count == 0)
                {
                    result.Add(curruentSubTree.Key);
                }

                foreach (var child in curruentSubTree.children)
                {
                    queue.Enqueue(child);
                }
            }

            return result;
        }

        public T GetDeepestKey()
        {
            throw new NotImplementedException();
        }

        public IEnumerable<T> GetLongestPath()
        {
            throw new NotImplementedException();
        }
    }
}
